+++ include/math/mat4.h
@@ -6,14 +6,14 @@
 
 class mat4 {
   public:
-    double m[4][4];
+    float m[4][4];
 
-    mat4(std::vector<std::vector<double>> m); 
+    mat4(std::vector<std::vector<float>> m); 
 
     vec4 operator*(const vec4 &right) const; 
 
-    static mat4 scale(double x, double y, double z);
-    static mat4 translate(double x, double y, double z);
+    static mat4 scale(float x, float y, float z);
+    static mat4 translate(float x, float y, float z);
     static mat4 identity();
 };
